requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-browser-reload==1.6.0
django-dotenv==1.4.2
idna==3.4
mercadopago==2.2.0
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
stripe==5.0.0
tzdata==2022.7
urllib3==1.26.13
